# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.CreateProductRequest:
    properties:
      category_id:
        type: string
      description:
        example: A very nice hat
        type: string
      images:
        items:
          type: string
        type: array
      name:
        example: Refined Bronze Hat
        maxLength: 100
        minLength: 3
        type: string
      sku:
        example: L1L-448
        type: string
      stock:
        example: 100
        minimum: 0
        type: integer
    required:
    - name
    - sku
    - stock
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: invalid request
        type: string
      message:
        example: failed request
        type: string
      status:
        example: false
        type: boolean
    type: object
...
  dto.ProductResponseWrapper:
    properties:
      data:
        $ref: '#/definitions/dto.ProductResponse'
      message:
        example: success get product
        type: string
      status:
        example: true
        type: boolean
    type: object
...
  dto.ProductsResponseWrapper:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.ProductListResponse'
        type: array
      message:
        example: success get products
        type: string
      meta: {}
      status:
        example: true
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@simponi.app
    name: Simponi Support
  description: |-
    REST API for Simponi - multi-store inventory and omnichannel management system.
    This API supports store management, product management, external product integration, vendor management, and role-based access control (RBAC).
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://simponi.app/terms/
  title: Simponi Backend API
  version: "1.0"
paths:
...
  /stores/{store_id}/products:
    get:
      consumes:
      - application/json
      description: 'Get paginated products in a store (Requires permission: GetProducts)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.ProductsResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get products in store
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: 'Create a new product inside a store (Requires permission: CreateProduct)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Create product request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.CreateProductRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/dto.ProductResponseWrapper'
        "400":
          description: Invalid input / UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict (SKU duplicate)
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create product
      tags:
      - Products
  /stores/{store_id}/products/{product_id}:
    delete:
      consumes:
      - application/json
      description: 'Delete product by ID (Requires permission: DeleteProductByStoreIDAndProductID)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.EmptySuccessResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: 'Get product detail by store ID and product ID (Requires permission:
        GetProductByStoreIDAndProductID)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.ProductResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get product detail
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: 'Update product information (Requires permission: UpdateProductByStoreIDAndProductID)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      - description: Update product request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateProductRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.ProductResponseWrapper'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product
      tags:
      - Products
  /stores/{store_id}/products/{product_id}/stock:
    patch:
      consumes:
      - application/json
      description: 'Update product stock with source tracking (Requires permission:
        UpdateStockByStoreIDAndProductID)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      - description: Stock update request
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateStockRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.EmptySuccessResponseWrapper'
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Product not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update product stock
      tags:
      - Products
  /stores/{store_id}/products/{product_id}/vendors/{vendor_id}:
    delete:
      consumes:
      - application/json
      description: 'Detach a vendor from a product inside a store (Requires permission:
        DetachVendorFromProduct)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      - description: Vendor ID (UUID)
        in: path
        name: vendor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.EmptySuccessResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Store/Product/Vendor relation not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Detach vendor from product
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: 'Attach a vendor to a product inside a store (Requires permission:
        AttachVendorToProduct)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      - description: Product ID (UUID)
        in: path
        name: product_id
        required: true
        type: string
      - description: Vendor ID (UUID)
        in: path
        name: vendor_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            $ref: '#/definitions/dto.EmptySuccessResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Store/Product/Vendor not found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Vendor already attached
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Attach vendor to product
      tags:
      - Products
  /stores/{store_id}/products/stats:
    get:
      consumes:
      - application/json
      description: 'Get aggregated product statistics in a store (Requires permission:
        GetProductStats)'
      parameters:
      - description: Store ID (UUID)
        in: path
        name: store_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/dto.ProductStatsResponseWrapper'
        "400":
          description: Invalid UUID
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get product statistics
      tags:
      - Products
...
securityDefinitions:
  BearerAuth:
    description: 'Use JWT token with format: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...